refactor metrics
